% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_initial_function.R
\name{encrypt_dashboard}
\alias{encrypt_dashboard}
\title{Encrypting dashboard}
\usage{
encrypt_dashboard()
}
\value{
Invisible NULL
}
\description{
Encrypting dashboard
}
\examples{
encrypt_dashboard()
}
