% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_initial_function.R
\name{get_central_time}
\alias{get_central_time}
\title{Get current time in Central Time Zone}
\usage{
get_central_time()
}
\value{
A character string with the formatted date and time
}
\description{
Returns the current date and time formatted in Central Time (US/Central)
with AM/PM format
}
\examples{
get_central_time()
}
