% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_initial_function.R
\name{get_enrollment_targets}
\alias{get_enrollment_targets}
\title{Get enrollment targets data}
\usage{
get_enrollment_targets()
}
\value{
A data frame with enrollment targets by date
}
\description{
Reads the enrollment targets CSV file and returns it as a data frame
}
\examples{
get_enrollment_targets()
}
