% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcal_api.R
\name{get_google_access_token}
\alias{get_google_access_token}
\title{Get Google Access Token}
\usage{
get_google_access_token(service_account)
}
\arguments{
\item{service_account}{List containing parsed service account JSON}
}
\value{
Character string containing the access token
}
\description{
Internal function to get an access token using service account credentials.
}
\keyword{internal}
