% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcal_api.R
\name{list_calendars}
\alias{list_calendars}
\title{List Google Calendars}
\usage{
list_calendars()
}
\value{
List containing calendar information
}
\description{
Lists all calendars the service account has access to.
}
\examples{
\dontrun{
calendars <- list_calendars()
}
}
