% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_login.R
\name{parse_csv}
\alias{parse_csv}
\title{Parse CSV Content}
\usage{
parse_csv(csv_content, save_path = NULL)
}
\arguments{
\item{csv_content}{Character string with the CSV content.}

\item{save_path}{Optional character string with file path to save the CSV.}
}
\value{
Data frame containing the CSV data.
}
\description{
Reads the CSV content into a data frame and optionally saves it to disk.
}
\keyword{internal}
