% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_initial_function.R
\name{run_initial_function}
\alias{run_initial_function}
\title{Initial function}
\usage{
run_initial_function(n)
}
\arguments{
\item{n}{A number}
}
\value{
Sum of n + n
}
\description{
Initial function
}
\examples{
run_initial_function(5)
run_initial_function(10)
run_initial_function(-3)
}
