% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_login.R
\name{test_abs_connection}
\alias{test_abs_connection}
\title{Test ABS Connection}
\usage{
test_abs_connection(base_url = "https://abs.la.utexas.edu", verbose = TRUE)
}
\arguments{
\item{base_url}{Character string with the base URL. Default is "https://abs.la.utexas.edu"}
}
\value{
Logical TRUE if server is reachable, FALSE otherwise
}
\description{
Simple function to test if the ABS server is reachable.
}
\examples{
\dontrun{
test_abs_connection()
}
}
